cdk/container/front-end-app/requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0
streamlit==1.29.0
boto3==1.33.11
streamlit-cognito-auth==1.2.0
langchain==0.1.7
traceloop-sdk==0.11.1
opentelemetry-instrumentation-bedrock
boto3